GreenAI_Waste_Segregation_Project/requirements.txt
streamlit
pandas
scikit-learn
joblib
reportlab
opencv-python
tensorflow
numpy
pillow
